remove bootstrap and applying Css and add some more JS Code
