--- src/LetsConnect/Connect.css
@@ -1,32 +1,0 @@
-.mainPage {
-  height: 90vh;
-  width: 65vh;
-  background-color: #ededc4;
-  margin-left: 30%;
-  padding-top: 150px;
-}
-
-.mb-3 {
-  padding-top: 10px;
-}
-
-.Query {
-  padding-top: 20px;
-}
-
-.query {
-  padding-top: 20px;
-  height: 70%;
-  width: 100%;
-}
-
-.LetsConnect {
-  display: flex;
-  justify-content: center;
-  padding-bottom: 10%;
-  font-weight: 900;
-}
-
-.btn {
-  margin: 3% 5% 2% 37%;
-}
